Better error messages and recovery
If MythX fails, it can return the empty string rather than valid data -
perhaps this is a Maestro bug? For now, we catch the string and turn it
into and empty object.

Give a better error message on a timeout. And we don't want tracebacks either here.

Start release notes.